# api/tests/postman/inventaio-core-api.openapi.yml
openapi: 3.1.0
info:
  title: InventAI/o API
  description: API del sistema inteligente de consulta y gestión logística para distribuidores de abarrotes.
  version: 1.0.0
servers:
- url: http://localhost:8000
  description: Docker Compose local
paths:
...
  /api/consulta/productos/{id_producto}/ventas:
    get:
      tags:
      - Consulta
      summary: Ventas de un producto en ventanas de 15 días hábiles
      description: 'INV-26: ventas de un producto en una sucursal física, en ventanas de 15 días hábiles (días con alguna
        venta en la red, como el modelo) que terminan en el último día hábil, para compararlas con POST /api/ml/predict. Fuente:
        dw.v_ventas_diarias_netas (sin devoluciones). Un día sin venta cuenta 0. Permisos de INV-25: gerente y admin_bodega
        deben enviar sucursal_id; admin_sucursal, la suya. Errores: 403 otra sucursal o la Bodega (admin_sucursal); 404 producto
        inexistente; 422 sin sucursal_id, sucursal inexistente, SIN_SUCURSAL, la Bodega Central (no vende) o ventanas fuera
        de 1-24.'
      operationId: ventas_producto_api_consulta_productos__id_producto__ventas_get
      security:
      - HTTPBearer: []
      parameters:
      - name: id_producto
        in: path
        required: true
        schema:
          type: integer
          title: Id Producto
        example: 171
      - name: sucursal_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: Sucursal física. Obligatoria para gerente y admin_bodega; admin_sucursal usa la suya.
          title: Sucursal Id
        description: Sucursal física. Obligatoria para gerente y admin_bodega; admin_sucursal usa la suya.
        example: 1
      - name: ventanas
        in: query
        required: false
        schema:
          type: integer
          maximum: 24
          minimum: 1
          description: Número de ventanas, de la más reciente hacia atrás
          default: 8
          title: Ventanas
        description: Número de ventanas, de la más reciente hacia atrás
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VentasProducto'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/consulta/inventario:
    get:
      tags:
      - Inventario
      summary: Stock actual con semáforo
      description: 'Retorna el inventario más reciente por producto-sucursal con semáforo (inconsistencia: stock negativo;
        si no, ok >7d, bajo 3-7d, critico <3d). Filtros: categoría, semáforo, búsqueda, sucursal. Cada fila trae tipo_ubicacion,
        stock_bodega (stock de la Bodega Central del mismo producto) y unidad (''kg'' si se vende por kilo; si no, ''unidad'').
        RBAC: admin_sucursal ve su sucursal y, con sucursal_id, el stock de la Bodega Central.'
      operationId: listar_inventario_api_consulta_inventario_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          default: 1
          title: Page
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Page Size
      - name: categoria
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filtrar por categoría
          title: Categoria
        description: Filtrar por categoría
        example: Huevos
      - name: semaforo
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: 'Filtrar: ok, bajo, critico, inconsistencia'
          title: Semaforo
        description: 'Filtrar: ok, bajo, critico, inconsistencia'
        example: critico
      - name: sucursal_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: 'Filtrar por sucursal. admin_sucursal: la suya o la Bodega Central; otra da 403. Un id inexistente
            o SIN_SUCURSAL da 422.'
          title: Sucursal Id
        description: 'Filtrar por sucursal. admin_sucursal: la suya o la Bodega Central; otra da 403. Un id inexistente o
          SIN_SUCURSAL da 422.'
        example: 1
      - name: busqueda
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Buscar en nombre producto
          title: Busqueda
        description: Buscar en nombre producto
        example: ARROZ
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventarioList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/consulta/inventario/detalle:
    get:
      tags:
      - Inventario
      summary: Historial 30 días de un producto
      description: 'Retorna el inventario actual y los últimos 30 días de un producto en una sucursal, con tipo_ubicacion
        y stock_bodega. Con la bodega real hay una sola foto: el historial trae un punto. RBAC: admin_sucursal, su sucursal
        o la Bodega Central.'
      operationId: detalle_inventario_api_consulta_inventario_detalle_get
      security:
      - HTTPBearer: []
      parameters:
      - name: id_producto
        in: query
        required: true
        schema:
          type: integer
          description: ID del producto
          title: Id Producto
        description: ID del producto
        example: 171
      - name: id_sucursal
        in: query
        required: true
        schema:
          type: integer
          description: ID de la sucursal
          title: Id Sucursal
        description: ID de la sucursal
        example: 1
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventarioDetalle'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/consulta/inventario/resumen:
    get:
      tags:
      - Inventario
      summary: Contadores por semáforo
      description: Retorna la cantidad de productos en cada estado del semáforo (ok, bajo, critico e inconsistencia), por
        sucursal y global. Con sucursal_id, solo esa ubicación.
      operationId: resumen_inventario_api_consulta_inventario_resumen_get
      security:
      - HTTPBearer: []
      parameters:
      - name: sucursal_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: 'Filtrar por sucursal. admin_sucursal: la suya o la Bodega Central; otra da 403. Un id inexistente
            o SIN_SUCURSAL da 422.'
          title: Sucursal Id
        description: 'Filtrar por sucursal. admin_sucursal: la suya o la Bodega Central; otra da 403. Un id inexistente o
          SIN_SUCURSAL da 422.'
        example: 1
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventarioResumenList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/consulta/inventario/valorizado:
    get:
      tags:
      - Inventario
      summary: Valor del stock por sucursal y categoría
      description: Retorna el valor monetario del inventario actual agrupado por sucursal y categoría. Con sucursal_id, solo
        esa ubicación.
      operationId: inventario_valorizado_api_consulta_inventario_valorizado_get
      security:
      - HTTPBearer: []
      parameters:
      - name: sucursal_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: 'Filtrar por sucursal. admin_sucursal: la suya o la Bodega Central; otra da 403. Un id inexistente
            o SIN_SUCURSAL da 422.'
          title: Sucursal Id
        description: 'Filtrar por sucursal. admin_sucursal: la suya o la Bodega Central; otra da 403. Un id inexistente o
          SIN_SUCURSAL da 422.'
        example: 1
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValorizadoList'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/reportes/ventas:
    get:
      tags:
      - Reportes
      summary: Ventas con filtros y agrupación
      description: 'Filtros: fecha_inicio, fecha_fin, sucursal, categoría. Agrupación: dia, semana, mes. datos_desde y datos_hasta:
        la primera y la última fecha con ventas.'
      operationId: ventas_api_reportes_ventas_get
      security:
      - HTTPBearer: []
      parameters:
      - name: fecha_inicio
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: YYYY-MM-DD
          title: Fecha Inicio
        description: YYYY-MM-DD
        example: '2025-12-01'
      - name: fecha_fin
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: YYYY-MM-DD
          title: Fecha Fin
        description: YYYY-MM-DD
        example: '2025-12-31'
      - name: sucursal_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          description: 'Filtrar por sucursal. admin_sucursal: solo la suya; otra da 403. Inexistente o SIN_SUCURSAL, 422'
          title: Sucursal Id
        description: 'Filtrar por sucursal. admin_sucursal: solo la suya; otra da 403. Inexistente o SIN_SUCURSAL, 422'
        example: 1
      - name: categoria
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Categoria
        example: Huevos
      - name: agrupacion
        in: query
        required: false
        schema:
          type: string
          description: dia, semana, mes
          default: dia
          title: Agrupacion
        description: dia, semana, mes
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VentasReporte'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    InventarioDetalle:
      properties:
        id_producto:
          type: integer
          title: Id Producto
        nombre_producto:
          type: string
          title: Nombre Producto
        categoria:
          type: string
          title: Categoria
        sucursal:
          type: string
          title: Sucursal
        id_sucursal:
          type: integer
          title: Id Sucursal
        tipo_ubicacion:
          type: string
          title: Tipo Ubicacion
        stock_actual:
          type: number
          title: Stock Actual
        stock_bodega:
          anyOf:
          - type: number
          - type: 'null'
          title: Stock Bodega
        stock_minimo:
          type: number
          title: Stock Minimo
        stock_maximo:
          type: number
          title: Stock Maximo
        punto_reorden:
          type: number
          title: Punto Reorden
        dias_cobertura:
          type: number
          title: Dias Cobertura
        semaforo:
          type: string
          title: Semaforo
        historial:
          items:
            $ref: '#/components/schemas/InventarioHistorialDia'
          type: array
          title: Historial
      type: object
      required:
      - id_producto
      - nombre_producto
      - categoria
      - sucursal
      - id_sucursal
      - tipo_ubicacion
      - stock_actual
      - stock_minimo
      - stock_maximo
      - punto_reorden
      - dias_cobertura
      - semaforo
      - historial
      title: InventarioDetalle
    InventarioHistorialDia:
      properties:
        fecha:
          type: string
          title: Fecha
        stock_disponible:
          type: number
          title: Stock Disponible
        dias_cobertura:
          type: number
          title: Dias Cobertura
        semaforo:
          type: string
          title: Semaforo
      type: object
      required:
      - fecha
      - stock_disponible
      - dias_cobertura
      - semaforo
      title: InventarioHistorialDia
    InventarioItem:
      properties:
        id_producto:
          type: integer
          title: Id Producto
        nombre_producto:
          type: string
          title: Nombre Producto
        categoria:
          type: string
          title: Categoria
        es_perecedero:
          type: boolean
          title: Es Perecedero
        unidad:
          type: string
          title: Unidad
        sucursal:
          type: string
          title: Sucursal
        id_sucursal:
          type: integer
          title: Id Sucursal
        tipo_ubicacion:
          type: string
          title: Tipo Ubicacion
        stock_disponible:
          type: number
          title: Stock Disponible
        stock_bodega:
          anyOf:
          - type: number
          - type: 'null'
          title: Stock Bodega
        stock_minimo:
          type: number
          title: Stock Minimo
        stock_maximo:
          type: number
          title: Stock Maximo
        punto_reorden:
          type: number
          title: Punto Reorden
        dias_cobertura:
          type: number
          title: Dias Cobertura
        semaforo:
          type: string
          title: Semaforo
        fecha:
          type: string
          title: Fecha
      type: object
      required:
      - id_producto
      - nombre_producto
      - categoria
      - es_perecedero
      - unidad
      - sucursal
      - id_sucursal
      - tipo_ubicacion
      - stock_disponible
      - stock_minimo
      - stock_maximo
      - punto_reorden
      - dias_cobertura
      - semaforo
      - fecha
      title: InventarioItem
    InventarioList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/InventarioItem'
          type: array
          title: Items
        total:
          type: integer
          title: Total
        page:
          type: integer
          title: Page
        page_size:
          type: integer
          title: Page Size
        pages:
          type: integer
          title: Pages
        fecha_inventario:
          type: string
          title: Fecha Inventario
      type: object
      required:
      - items
      - total
      - page
      - page_size
      - pages
      - fecha_inventario
      title: InventarioList
    InventarioResumen:
      properties:
        sucursal:
          anyOf:
          - type: string
          - type: 'null'
          title: Sucursal
        id_sucursal:
          anyOf:
          - type: integer
          - type: 'null'
          title: Id Sucursal
        tipo_ubicacion:
          anyOf:
          - type: string
          - type: 'null'
          title: Tipo Ubicacion
        contadores:
          $ref: '#/components/schemas/SemaforoContador'
        fecha_inventario:
          type: string
          title: Fecha Inventario
      type: object
      required:
      - contadores
      - fecha_inventario
      title: InventarioResumen
    InventarioResumenList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/InventarioResumen'
          type: array
          title: Items
        global_:
          $ref: '#/components/schemas/SemaforoContador'
        fecha_inventario:
          type: string
          title: Fecha Inventario
      type: object
      required:
      - items
      - global_
      - fecha_inventario
      title: InventarioResumenList
...
    SemaforoContador:
      properties:
        ok:
          type: integer
          title: Ok
        bajo:
          type: integer
          title: Bajo
        critico:
          type: integer
          title: Critico
        inconsistencia:
          type: integer
          title: Inconsistencia
        total:
          type: integer
          title: Total
      type: object
      required:
      - ok
      - bajo
      - critico
      - inconsistencia
      - total
      title: SemaforoContador
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    ValorizadoList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ValorizadoItem'
          type: array
          title: Items
        total_valor:
          type: number
          title: Total Valor
        fecha_inventario:
          type: string
          title: Fecha Inventario
      type: object
      required:
      - items
      - total_valor
      - fecha_inventario
      title: ValorizadoList
...
    VentanaVentas:
      properties:
        desde:
          type: string
          format: date
          title: Desde
          description: Primer día hábil de la ventana.
        hasta:
          type: string
          format: date
          title: Hasta
          description: Último día hábil de la ventana.
        unidades:
          type: number
          title: Unidades
          description: Unidades vendidas en los días hábiles de la ventana (0 si no hubo venta).
      type: object
      required:
      - desde
      - hasta
      - unidades
      title: VentanaVentas
...
    VentasProducto:
      properties:
        id_producto:
          type: integer
          title: Id Producto
        codigo_item:
          type: string
          title: Codigo Item
        nombre_producto:
          type: string
          title: Nombre Producto
        categoria:
          type: string
          title: Categoria
        unidad:
          type: string
          title: Unidad
          description: 'Unidad de las ventanas: ''kg'' si se vende por kilo; si no, ''unidad''.'
        id_sucursal:
          type: integer
          title: Id Sucursal
        sucursal:
          type: string
          title: Sucursal
        horizonte_dias_habiles:
          type: integer
          title: Horizonte Dias Habiles
          description: 'Días hábiles de cada ventana: 15, el horizonte del pronóstico.'
        fecha_fin:
          type: string
          format: date
          title: Fecha Fin
          description: Último día hábil de la red; la última ventana termina ahí.
        ventanas:
          items:
            $ref: '#/components/schemas/VentanaVentas'
          type: array
          title: Ventanas
          description: En orden cronológico.
      type: object
      required:
      - id_producto
      - codigo_item
      - nombre_producto
      - categoria
      - unidad
      - id_sucursal
      - sucursal
      - horizonte_dias_habiles
      - fecha_fin
      - ventanas
      title: VentasProducto
    VentasReporte:
      properties:
        items:
          items:
            $ref: '#/components/schemas/VentaPeriodo'
          type: array
          title: Items
        total_cantidad:
          type: number
          title: Total Cantidad
        total_valor:
          type: number
          title: Total Valor
        total_margen:
          type: number
          title: Total Margen
        agrupacion:
          type: string
          title: Agrupacion
        fecha_inicio:
          type: string
          title: Fecha Inicio
        fecha_fin:
          type: string
          title: Fecha Fin
        datos_desde:
          type: string
          title: Datos Desde
        datos_hasta:
          type: string
          title: Datos Hasta
      type: object
      required:
      - items
      - total_cantidad
      - total_valor
      - total_margen
      - agrupacion
      - fecha_inicio
      - fecha_fin
      - datos_desde
      - datos_hasta
      title: VentasReporte
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer